mcp-servers/monitor/requirements.txt
# MCP Monitor Server (EC2 per-service venv)
mcp==1.26.0
httpx==0.28.1
python-dotenv==1.2.2
